Refactoring, alternate fix
